Out-of-tree Topology-aware scheduler plugin
As part of the enablement of Topology-aware scheduling in kubernetes,
the scheduler plugin has been proposed as in-tree.

KEP: https://github.com/kubernetes/enhancements/pull/1858
PR: https://github.com/kubernetes/kubernetes/pull/90708

To enable faster development velocity, testing and community adoption
we are also packaging as an out of tree scheduler plugin. This out-of-tree
implementation is based on the above PR and KEP.